+++ src/main/java/com/codeaddi/mrc_resources/controller/service/ResourceService.java
@@ -1,0 +1,4 @@
+package com.codeaddi.mrc_resources.controller.service;
+
+public class ResourceService {
+}
